Added marker data to frames and keyframes & added viewer marker visuals
Each marker is made up of 4 points. These points have been undistorted
and can be either used as another source of mapping data or matched with
orb features in the frame ( e.g. using nearest neighbour). The markers
return distance and pose of the camera with respect to the marker board
plane